test(coverage): expand coverage metrics, add LFS mock tests, and resolve sanitizer warnings
- Raised line coverage to 99.5% and function coverage to 100% (595 tests passing).
- Added branch coverage to Makefile.unity and local config, raising overall branch coverage to 75.8%.
- Added LittleFS mock tests and port layer definitions (lfs.h, test_lfs.c).
- Fixed critical AddressSanitizer and UndefinedBehaviorSanitizer warnings (stack-buffer-overflows in JSON parser, global-buffer-overflows in CBOR reader/mock tests, undefined negative left shifts, null-pointer memcpy parameters).
- Added static stack usage and runtime sanitizer guides to testing.md docs.

+++ src/syntropic/util/syn_qmath.h
@@ -41,7 +41,7 @@ typedef int32_t q16_t;
 #define Q16_FROM_INT(n)       ((q16_t)((int32_t)(n) * Q16_ONE))
 
 /** Fraction to Q16: Q16_FROM_FRAC(1, 3) ≈ 0.333. */
-#define Q16_FROM_FRAC(num, den) ((q16_t)(((int64_t)(num) << Q16_SHIFT) / (den)))
+#define Q16_FROM_FRAC(num, den) ((q16_t)(((int64_t)((uint64_t)(int64_t)(num) << Q16_SHIFT)) / (den)))
 
 /** Float literal to Q16 (compile-time only, avoid at runtime). */
 #define Q16_FROM_FLOAT(f)     ((q16_t)((f) * (1L << Q16_SHIFT)))
@@ -76,7 +76,7 @@ static inline q16_t q16_mul(q16_t a, q16_t b)
  */
 static inline q16_t q16_div(q16_t a, q16_t b)
 {
-    return (q16_t)(((int64_t)a << Q16_SHIFT) / b);
+    return (q16_t)(((int64_t)((uint64_t)(int64_t)a << Q16_SHIFT)) / b);
 }
 
 /**
